some warning fixes

+++ engine/utils/vec.cpp
@@ -471,24 +471,24 @@ float lengthSqr(const vec4 &v){
 }
 
 vec2 floor(const vec2 &v){
-    return vec2(floor(v.x), floor(v.y));
+    return vec2(floorf(v.x), floorf(v.y));
 }
 
 vec3 floor(const vec3 &v){
-    return vec3(floor(v.x), floor(v.y), floor(v.z));
+    return vec3(floorf(v.x), floorf(v.y), floorf(v.z));
 }
 vec4 floor(const vec4 &v){
-    return vec4(floor(v.x), floor(v.y), floor(v.z), floor(v.w));
+    return vec4(floorf(v.x), floorf(v.y), floorf(v.z), floorf(v.w));
 }
 
 vec2 frac(const vec2 &v){
-    return vec2(v.x - floor(v.x), v.y - floor(v.y));
+    return vec2(v.x - floorf(v.x), v.y - floorf(v.y));
 }
 vec3 frac(const vec3 &v){
-    return vec3(v.x - floor(v.x), v.y - floor(v.y), v.z - floor(v.z));
+    return vec3(v.x - floorf(v.x), v.y - floorf(v.y), v.z - floorf(v.z));
 }
 vec4 frac(const vec4 &v){
-    return vec4(v.x - floor(v.x), v.y - floor(v.y), v.z - floor(v.z), v.w - floor(v.w));
+    return vec4(v.x - floorf(v.x), v.y - floorf(v.y), v.z - floorf(v.z), v.w - floorf(v.w));
 }
 
 vec3 cross(const vec3 &u, const vec3 &v){
